Rename zgs module to zonegroupstate
